Removed INVERSE_1D and added MULTIPLE_FORWARD
The inverse_1d function is based on the original 3D STSN model. I have removed it so that a dispersion based STSN model can be used in the future instead. For now, I've added a multiple_forward model that can handle Lorentz model dispersion with multiple resonance frequencies. For now, the code is only stable for single resonances but will hopefully be ready for multiple resonances soon.